+++ argilla/src/argilla/_models/_record/_metadata.py
@@ -12,16 +12,13 @@
 # See the License for the specific language governing permissions and
 # limitations under the License.
 
-from typing import Annotated, Union, List
+from typing import Annotated, Any
 
 from pydantic import BaseModel
-
-
-MetadataValue = Annotated[Union[str, List[str], float, int, None], "The value of the metadata field dictionary"]
 
 
 class MetadataModel(BaseModel):
     """Schema for the metadata of a `Dataset`"""
 
     name: Annotated[str, "The name of the metadata field or key in the metadata dictionary"]
-    value: MetadataValue
+    value: Any
